+++ tests/testthat/test-atc.R
@@ -1,8 +1,7 @@
 context("atc.R")
 
 test_that("atc_property works", {
-  skip_on_travis() # relies on internet connection of server, don't test
-
+  skip_on_cran() # relies on internet connection of server, don't test
   skip_on_appveyor() # security error on AppVeyor
 
   if (!is.null(curl::nslookup("www.whocc.no", error = FALSE))) {